AI Healthcare/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
pydicom
albumentations
Pillow
grad-cam
numpy
timm
